+++ physical_ai_tools/CHANGELOG.rst
@@ -2,6 +2,22 @@
 Changelog for package physical_ai_tools
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
+0.7.2 (2025-12-01)
+------------------
+* Fixed an issue where the task_index was being merged based on the first episode when merging episodes in the *.parquet data.
+* Contributors: Dongyun Kim
+
+0.7.1 (2025-11-28)
+------------------
+* Updated physical_ai_server launch file to use direct rosbridge websocket node instantiation
+* Contributors: Kiwoong Park
+
+0.7.0 (2025-11-21)
+------------------
+* Added rosbag_recorder package
+* Added rosbag2 recording support when collecting LeRobot datasets
+* Contributors: Woojin Wie, Kiwoong Park
+
 0.6.13 (2025-10-27)
 ------------------
 * Fixed physical_ai_server crash when querying user ID without locally registered HuggingFace token